error: implicit conversion loses integer precision: 'long' to 'int' [-Werror,-Wshorten-64-to-32]

+++ libc/test/src/sys/socket/linux/sendmsg_recvmsg_test.cpp
@@ -63,7 +63,7 @@ TEST_F(LlvmLibcSendMsgRecvMsgTest, SucceedsWithSocketPair) {
   recv_message.msg_controllen = 0;
   recv_message.msg_flags = 0;
 
-  ASSERT_THAT(LIBC_NAMESPACE::recvmsg(sockpair[1], &recv_message, 0),
+  ASSERT_THAT((int)LIBC_NAMESPACE::recvmsg(sockpair[1], &recv_message, 0),
               Succeeds(static_cast<ssize_t>(MESSAGE_LEN)));
 
   ASSERT_STREQ(buffer, TEST_MESSAGE);
@@ -110,5 +110,5 @@ TEST_F(LlvmLibcSendMsgRecvMsgTest, RecvFails) {
   recv_message.msg_controllen = 0;
   recv_message.msg_flags = 0;
 
-  ASSERT_THAT(LIBC_NAMESPACE::recvmsg(-1, &recv_message, 0), Fails(EBADF));
+  ASSERT_THAT((int)LIBC_NAMESPACE::recvmsg(-1, &recv_message, 0), Fails(EBADF));
 }
